% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_metadata}
\alias{download_metadata}
\title{Download the metadata found on the \href{https://voteview.polisci.ucla.edu/data}{VoteView Data page}
for members, rollcalls, or parties.}
\usage{
download_metadata(type, chamber = "both", congress = "all",
  usetemp = TRUE)
}
\arguments{
\item{type}{A string that is either 'member', 'rollcall', or 'party'.}

\item{chamber}{A string that defines the chamber for which data should be
returned, either 'both', 'house', or 'senate'. Defaults to 'both' and is
ignored if the \code{type} is 'party'.}

\item{congress}{Either 'all' or a numeric vector that denotes the congresses
for which data can be returned. Defaults to 'all'}
}
\value{
A data frame with all relevant metadata.
}
\description{
Download the metadata found on the \href{https://voteview.polisci.ucla.edu/data}{VoteView Data page}
for members, rollcalls, or parties.
}
