% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollcallUtilities.R
\name{\%+\%}
\alias{\%+\%}
\title{Full outer join of two rollcall objects}
\usage{
rc1 \%+\% rc2
}
\arguments{
\item{rc1}{The first rollcall object.}

\item{rc2}{The second rollcall object.}
}
\value{
The union of the two rollcall objects.
}
\description{
This function takes two rollcall objects as its parameters and performs a
full outer join on them, returning a new rollcall object that is the union
of the two.
}
\examples{

## Search for rollcalls about Rhodesia in the 96th Congress
res <- voteview_search("Rhodesia", congress = c(96))

rc1 <- voteview_download(res$id)
summary(rc1)

## Search for rollcalls about Rhodesia in the 90th Congress
res <- voteview_search("Rhodesia", congress = c(90))

rc2 <- voteview_download(res$id)
summary(rc2)

## Union of the two results
rc <- rc1 \%+\% rc2

}
