% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{jlist2df}
\alias{jlist2df}
\title{Transform vector of lists to data frame (DEPRECATED)}
\usage{
jlist2df(rcs, ordercols = NULL)
}
\arguments{
\item{rcs}{A vector of lists that is built in a json download
where each list corresponds to a roll call or a member depending on the
function that called it.}

\item{ordercols}{A order of columns so that the returned data.frame is easy
to read.}
}
\description{
This function has been deprecated in favor of using fromJSON in jsonlite. This is a helper function to transform the vector of lists that were
constructed by \code{voteview_search} and \code{member_search} in to a data frame.
This function should probably not be called
by itself. See \code{voteview_search} and \code{member_search} for more information.
}
\seealso{
'\link{voteview_search}', '\link{member_search}'.
}
