% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollcallUtilities.R
\name{melt_rollcall}
\alias{melt_rollcall}
\title{Melt rollcall object to long rollcall data frame}
\usage{
melt_rollcall(rc, keepvote = rownames(rc$vote.data), legiscols = NULL,
  votecols = NULL, dropNotInLegis = TRUE)
}
\arguments{
\item{rc}{A rollcall object.}

\item{keepvote}{A character vector of vote names to keep. Defaults to all.}

\item{legiscols}{A character vector of column names from \code{legis.data} to
include. Defaults to all columns.}

\item{votecols}{A character vector of column names from \code{vote.data} to
include. Defaults to all columns.}

\item{dropNotInLegis}{Boolean to drop votes that were "notInLegis". Defaults
to TRUE.}
}
\value{
A data.frame with the following columns:
\itemize{
\item{\code{legisid} }{Legislator names from rollcall object}
\item{\code{voteid} }{Vote names from rollcall object}
\item{\code{legiscols} }{All columns selected from \code{legis.data}}
\item{\code{votecols} }{All columns selected from \code{vote.data}}
\item{\code{vote} }{Numeric vote for legislator-rollcall pair}
}
}
\description{
This function takes a rollcall object and melts it into a long format, merging
attributes of the rollcalls and legislators. This can be used for legislator-rollcall
analyses.
}
\details{
This function works for any \code{rollcall} object that meets the following
two criteria. The \code{vote.names}, or the row names of the \code{vote.data}
matrix within the \code{rollcall} object, must match the column names of the 
\code{votes} matrix. Similarly, the 
\code{legis.names}, or the row names of the \code{legis.data} matrix within 
the \code{rollcall} object, must match the row names of the \code{votes} matrix.

This will work for all \code{rollcall} objects downloaded from the voteview database.
If \code{legis.long.dynamic} and \code{votes.long} are included, meaning \code{keeplong == TRUE},
then the melted data frame will be unique by internal id and roll call. Otherwise
the melted data frame will be unique by icpsr and roll call.
}
\examples{

## Search for example roll calls
res <- voteview_search("Rhodesia")

## Download some ids
rc <- voteview_download(res$id[1:10])

## Create long rollcall object, wihtout long description fields
rclong <- melt_rollcall(rc, votecols = c("chamber", "congress"))

}
\seealso{
'\link[pscl]{rollcall}'.
}

