% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteview.R
\name{member_search}
\alias{member_search}
\title{Query the Voteview Database for Members}
\usage{
member_search(name = NULL, icpsr = NULL, state = NULL, congress = NULL,
  cqlabel = NULL, chamber = NULL)
}
\arguments{
\item{name}{The name you wish to search}
}
\value{
A data.frame with data for members of Congress. Each row is a unique member in our database.
}
\description{
This function is under construction. Use with caution.
}
\details{
This function searches the database for specific members.
}
\examples{

## Search for obama
res <- member_search("obama")

}
\seealso{
'\link{voteview_search}'.
}

