\name{rvoteview-package}
\alias{rvoteview-package}
\alias{rvoteview}
\docType{package}
\title{
Run queries within voteview
}
\description{
Package provides functions to download, read, and search Roll Call votes from the U.S. Congress.
}
\details{
\tabular{ll}{
Package: \tab rvoteview\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-24\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Jeff Lewis

Felipe Nunes <felipenunes@ucla.edu>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
