% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Data of state codes and names}
\format{An object of class \code{data.frame} with 51 rows and 5 columns.}
\usage{
states
}
\description{
Dataset containing the ICPSR state codes, mail codes, and the state names.
}
\keyword{datasets}
