% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{votelist2voteview}
\alias{votelist2voteview}
\title{Internal function to build a voteview object from downloaded votelist}
\usage{
votelist2voteview(dat)
}
\description{
Internal function to build a voteview object from downloaded votelist
}

