% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteview.R
\name{voteview.download.json}
\alias{voteview.download.json}
\title{Download voteview object from voteview server}
\usage{
voteview.download.json(ids)
}
\arguments{
\item{ids}{A string or vector of strings, where each string is a roll call
ID, such as "S1110326". These ids can be found using the
\code{voteview.search} function, if they are not already known.}
}
\value{
A single string of JSON code.
}
\description{
This internal function takes roll call IDs and returns a JSON string that contains data on the
description, date, all individual legislator's votes, and more on each roll
call.
}
\examples{

## Search for sample roll calls
res <- voteview.search("Iraq")
  
## Use the ids from that search to create a json string
json <- voteview.download.json(res$ids)
}
\seealso{
'\link{voteview.download}','\link{voteview.search}'.
}

