% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteview.R
\name{voteview_search}
\alias{voteview_search}
\title{Query the Voteview Database with a String}
\usage{
voteview_search(alltext = NULL, startdate = NULL, enddate = NULL,
  session = NULL, chamber = NULL, maxsupport = NULL, minsupport = NULL,
  query = NULL)
}
\arguments{
\item{alltext}{A string or character vector that searches all text fields in
the voteview datase. All words are searched independently and joined with 
an OR statement. Case-insensitive.}

\item{startdate}{A string of the format \code{"yyyy-mm-dd"} that is the
earliest possible date to search for a roll call.}

\item{enddate}{A string of the format
"yyyy-mm-dd" that is the earliest possible date to search for a roll
call.}

\item{session}{A numeric vector of the sessions of congress to constrain the search to. The default is all sessions.}

\item{chamber}{A string in \code{c("House", "Senate")}. The default
NULL value returns results from both chambers of congress.}

\item{maxsupport}{NOT CURRENTLY SUPPORTED. Support is the share of Yea votes 
over total Yea and Nay votes. \code{maxsupport} is a number specifying the 
maximum support allowed for returned votes.}

\item{minsupport}{NOT CURRENTLY SUPPORTED. A number specifying the minimum 
support allowed for returned votes.}

\item{query}{NOT CURRENTLY SUPPORTED. A string that can specify a more 
complex search of all of the text fields. See examples for usage and syntax.
 Can only be used if \code{alltext} is not specified.}
}
\value{
A data.frame with the following columns: 
\itemize{
\item{\code{description} }{Official description of the bill.}
\item{\code{shortdescription} }{A short description of the bill.}
\item{\code{date} }{The date the roll call was held, in string "yyyy-mm-dd"
format.}
\item{\code{bill} }{Bill name abbreviation.}
\item{\code{chamber} }{The chamber the roll call was held in. Either "House"
or "Senate"} 
\item{\code{session} }{The session of congress the roll call was held in.}
\item{\code{rollnumber} }{The roll call number of the vote.}
\item{\code{yea} }{The total number of 'Yea' votes.}
\item{\code{nay} }{The total number of 'Nay' votes.}
\item{\code{support} }{Percent of 'Yea' votes out of all 'Yea' or 'Nay' 
votes (excludes absent).}
\item{\code{id} }{Unique identifier for the roll call that allows
\code{voteview} to query the individual congress people's votes and other
details.}
}
}
\description{
Searches the Voteview database with a string and returns a data frame with
bill IDs, summary vote statistics, and other identifying information.
}
\details{
Each of the fields is currently joined together with an AND statement.

\code{query} uses the following general syntax, \code{field:query AND (query OR query)}. For example, if you wanted to find votes with "war" and either "iraq" or "afghanistan" in any text field, you could set the query to be \code{"alltext:war AND (iraq OR afghanistan)"}. If you wanted to do the same query but only return the votes with "defense" in the description field, the query would become \code{"alltext:war AND (iraq OR afghanistan) description:defense"}

The fields that can be searched with text are \code{code.Clausen}, \code{code.Peltzman}, \code{code.Issue}, \code{description}, \code{shortdescription}, \code{bill}, and \code{alltext}. The fields that can be searched by number are \code{session}, \code{yea}, \code{nay}, and \code{support}.
}
\examples{

## Search for example roll calls
res <- voteview_search("Iraq")
res

\dontrun{
## Search for votes with a start date
res <- voteview_search("Iraq", startdate = "2005-01-01")
res
 
## Search for votes with an end date in just the house
res <- voteview_search("Iraq", startdate = "2005-01-01", chamber = "House")
res

#' ## Search for votes with an end date in just the house in 112 session
res <- voteview_search("Iraq", startdate = "2005-01-01", session = 112, chamber = "House")
res
}
}
\seealso{
'\link{voteview_download}'.
}

