% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{build_votelist}
\alias{build_votelist}
\title{Internal function to get rollcalls and build voteview list}
\usage{
build_votelist(votelist, ids, perrequest)
}
\description{
Internal function to get rollcalls and build voteview list
}

