% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteview.R
\name{vlist2df}
\alias{vlist2df}
\title{Transform vector of lists to data frame}
\usage{
vlist2df(rcs)
}
\arguments{
\item{rcs}{A vector of lists that is built within \code{voteview.search},
where each list corresponds to a roll call.}
}
\description{
This is a helper function to transform the vector of lists that were
constructed by \code{voteview.search} in to a data frame that
\code{voteview.search} returns. This function should probably not be called
by itself. See \code{voteview.search} for more information.
}
\seealso{
'\link{voteview.search}'.
}

