% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteview.R
\name{voteview2rollcall}
\alias{voteview2rollcall}
\title{Internal function that transforms Voteview object to a rollcall object}
\usage{
voteview2rollcall(data, keeplong = T)
}
\description{
Internal function that transforms Voteview object to a rollcall object
}

